+++ sdk/hardwaresecuritymodules/arm-hardwaresecuritymodules/samples/v3-beta/javascript/cloudHsmClusterRestoreStatusGetSample.js
@@ -5,24 +5,25 @@ const { AzureDedicatedHSMResourceProvider } = require("@azure/arm-hardwaresecuri
 const { DefaultAzureCredential } = require("@azure/identity");
 
 /**
- * This sample demonstrates how to gets the backup operation status of the specified Cloud HSM Cluster
+ * This sample demonstrates how to gets the restore operation status of the specified Cloud HSM Cluster
  *
- * @summary gets the backup operation status of the specified Cloud HSM Cluster
- * x-ms-original-file: 2025-03-31/CloudHsmCluster_Backup_Pending_MaximumSet_Gen.json
+ * @summary gets the restore operation status of the specified Cloud HSM Cluster
+ * x-ms-original-file: 2025-03-31/CloudHsmCluster_Restore_Pending_MaximumSet_Gen.json
  */
-async function cloudHsmClusterGetBackupStatusMaximumSetGen() {
+async function cloudHsmClusterGetRestoreStatusMaximumSetGen() {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "00000000-0000-0000-0000-000000000000";
   const client = new AzureDedicatedHSMResourceProvider(credential, subscriptionId);
-  await client.cloudHsmClusterBackupStatus.get(
+  const result = await client.cloudHsmClusterRestoreStatus.get(
     "rgcloudhsm",
     "chsm1",
     "572a45927fc240e1ac075de27371680b",
   );
+  console.log(result);
 }
 
 async function main() {
-  await cloudHsmClusterGetBackupStatusMaximumSetGen();
+  await cloudHsmClusterGetRestoreStatusMaximumSetGen();
 }
 
 main().catch(console.error);