+++ sdk/hardwaresecuritymodules/arm-hardwaresecuritymodules/samples/v3-beta/typescript/src/cloudHsmClusterBackupStatusGetSample.ts
@@ -14,11 +14,12 @@ async function cloudHsmClusterGetBackupStatusMaximumSetGen(): Promise<void> {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "00000000-0000-0000-0000-000000000000";
   const client = new AzureDedicatedHSMResourceProvider(credential, subscriptionId);
-  await client.cloudHsmClusterBackupStatus.get(
+  const result = await client.cloudHsmClusterBackupStatus.get(
     "rgcloudhsm",
     "chsm1",
     "572a45927fc240e1ac075de27371680b",
   );
+  console.log(result);
 }
 
 async function main(): Promise<void> {