+++ sdk/hardwaresecuritymodules/arm-hardwaresecuritymodules/src/classic/cloudHsmClusterRestoreStatus/index.ts
@@ -2,9 +2,9 @@
 // Licensed under the MIT License.
 
 import { AzureDedicatedHSMResourceProviderContext } from "../../api/azureDedicatedHSMResourceProviderContext.js";
+import { RestoreResult } from "../../models/models.js";
 import { CloudHsmClusterRestoreStatusGetOptionalParams } from "../../api/cloudHsmClusterRestoreStatus/options.js";
 import { get } from "../../api/cloudHsmClusterRestoreStatus/operations.js";
-import { PollerLike, OperationState } from "@azure/core-lro";
 
 /** Interface representing a CloudHsmClusterRestoreStatus operations. */
 export interface CloudHsmClusterRestoreStatusOperations {
@@ -14,7 +14,7 @@ export interface CloudHsmClusterRestoreStatusOperations {
     cloudHsmClusterName: string,
     jobId: string,
     options?: CloudHsmClusterRestoreStatusGetOptionalParams,
-  ) => PollerLike<OperationState<void>, void>;
+  ) => Promise<RestoreResult | null>;
 }
 
 function _getCloudHsmClusterRestoreStatus(context: AzureDedicatedHSMResourceProviderContext) {